frontend completed

+++ frontend/src/components/WorkoutDetails.js
@@ -1,17 +1,30 @@
+//In this component we try to DELETE a workout
+
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
+import { useAuthContext } from '../hooks/useAuthContext'
 //date fns
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
+
+
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext()
 
+    
   const handleClick = async () => {
+    if (!user) {
+      return
+    }
     const response = await fetch('/api/workouts/' + workout._id, {
       method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     });
     const json = await response.json();
 
     if (response.ok) {
-      dispatch({ type: "DELETE_WORKOUTS", payload: json });
+      dispatch({ type: "DELETE_WORKOUTS", payload: workout._id });
     }
   };
 
@@ -35,4 +48,4 @@ const WorkoutDetails = ({ workout }) => {
     </div>
   );
 };
-export default WorkoutDetails;
+export default WorkoutDetails
